+++ examples/23-bundlerApiKey.ts
@@ -1,12 +1,15 @@
-import { PrimeSdk } from '../src';
+import { EtherspotBundler, PrimeSdk } from '../src';
 import * as dotenv from 'dotenv';
 
 dotenv.config();
 
 
 async function main() {
+  const etherspotBundlerApiKey = '';
   // initializating sdk...
-  const primeSdk = new PrimeSdk({ privateKey: process.env.WALLET_PRIVATE_KEY }, { chainId: Number(process.env.CHAIN_ID), projectKey: 'public-prime-testnet-key', etherspotBundlerApiKey: '' })
+  const primeSdk = new PrimeSdk({ privateKey: process.env.WALLET_PRIVATE_KEY }, { chainId: Number(process.env.CHAIN_ID), projectKey: 'public-prime-testnet-key', 
+    bundlerProvider: new EtherspotBundler(Number(process.env.CHAIN_ID), etherspotBundlerApiKey) 
+  })
 
   // get EtherspotWallet address...
   const address: string = await primeSdk.getCounterFactualAddress();