+++ src/sdk/bundler/providers/GenericBundler.ts
@@ -1,0 +1,9 @@
+import { BundlerProvider } from "../interface";
+
+
+export class GenericBundler implements BundlerProvider {
+  readonly url: string;
+  constructor(bundlerUrl: string) {
+    this.url = bundlerUrl;
+  }
+}