+++ src/sdk/interfaces.ts
@@ -1,3 +1,4 @@
+import { BundlerProviderLike } from './bundler';
 import { StateStorage } from './state';
 
 export interface PaymasterApi {
@@ -14,8 +15,8 @@ export enum Factory {
 export interface SdkOptions {
   chainId: number;
   projectKey: string;
+  bundlerProvider: BundlerProviderLike;
   stateStorage?: StateStorage;
-  bundlerRpcUrl?: string;
   rpcProviderUrl?: string;
   graphqlEndpoint?: string;
   etherspotBundlerApiKey?: string;